Consolidate more wallet-creation functions
